requirements.txt
selenium==4.1.0
Pillow==9.0.0